Updated to be organised into partial and completed concurrent for now.
