+++ Group.NET/Group/Group.Hierarchy.cs
@@ -1,0 +1,13 @@
+﻿using Group.NET.Group.Interfaces;
+using System.Collections.Concurrent;
+using System.Collections.Immutable;
+
+namespace Group.NET
+{
+    public partial class Group<TKey, TValue>
+        where TKey : IEquatable<TKey>
+    {
+        private readonly IDictionary<TKey, Group<TKey, TValue>> _childrenGroups;
+
+    }
+}
