--- Group.NET/IGroupHierarchyReadOnly.cs
@@ -1,9 +1,0 @@
-﻿namespace Group.NET
-{
-    public interface IGroupHierarchyReadOnly<TKey, TValue>
-        where TKey: IEquatable<TKey>
-    {
-        IEnumerable<TKey> GetKeysForChildrenGroups();
-        bool IsRootGroup { get; }
-    }
-}
